Update patch set 1
Patch Set 1:

Follow up to what has been said here: http://review.cyanogenmod.org/52170

In this workaround as it is now, I don't like the fact that the proximity sensor would be enabled virtually inifinitely since boot till the first phone call end. While it may not affect battery too much, every mA counts.

What seems to happen at least on Photon Q is that the proximity sensor is not working correctly for some time after it has been enabled for the first time. Therefore the issue usually affects only the first call after boot. (The sensor is normally first enabled by the first call, disabled again at the call end). After some time of the sensor being on, it starts to work fine, no matter how many times it's been toggled and how long time it spent in disabled state inbetween the toggles.

The initial state of prox_enable is 0 (disabled). So the echo 0 > .../prox_enable has no effect, what counts here is that the proximity sensor is enabled early by the second echo 1 > .../prox_enable .

What works for me on Q is:

echo 1 > .../prox_enable
sleep 10
echo 0 > .../prox_enable

This is better in regard to battery as it does not leave the sensor enabled infinitely, yet it fixes the issue for me. 10s is just a number I've tried and found working. It should be adjusted to an as low value as possible while long enough to fix it for all devices.

@Ricky: Can you test this approach on Atrix HD? Thanks.

Patch-set: 1