Update patch set 1
Patch Set 1:

Another note: this patch disables the use of the back mic located near camera and enables only the front mic. This is fine if you want to capture the narration of the cameraman, but the usual use case is rather to capture the sounds that happen in front of the camera.

Patch-set: 1