Update patch set 2
Patch Set 2:

Emerson: I'm trying to analyze the issue by comparing the differences between the stock and our audio capture behavior, but I haven't managed to get the right exact debug data so far. But I suppose that the difference is in an additional audio processing done (or triggered) by the stock audio HAL that's missing in our open source solution.

Patch-set: 2